backend/.agents/skills-custom/book-review/scripts/requirements.txt
openai>=1.30.0
pymysql==1.1.0
pydantic==2.5.0
python-dotenv==1.0.0
pypdf==4.0.1
pillow==10.1.0
jieba>=0.42.1
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
minio>=7.2.0
